+++ jabgui/src/main/java/org/jabref/gui/libraryproperties/LibraryPropertiesView.java
@@ -9,10 +9,10 @@ import javafx.scene.control.TabPane;
 import org.jabref.gui.StateManager;
 import org.jabref.gui.util.BaseDialog;
 import org.jabref.gui.util.ControlHelper;
+import org.jabref.gui.util.ViewLoader;
 import org.jabref.logic.l10n.Localization;
 import org.jabref.model.database.BibDatabaseContext;
 
-import com.airhacks.afterburner.views.ViewLoader;
 import jakarta.inject.Inject;
 
 public class LibraryPropertiesView extends BaseDialog<LibraryPropertiesViewModel> {