+++ jabgui/src/main/java/org/jabref/gui/maintable/MainTableColumnModel.java
@@ -14,6 +14,7 @@ import javafx.beans.property.StringProperty;
 import javafx.scene.control.TableColumn;
 
 import org.jabref.gui.util.FieldsUtil;
+import org.jabref.injection.Injector;
 import org.jabref.logic.l10n.Localization;
 import org.jabref.model.entry.field.FieldFactory;
 import org.jabref.model.metadata.SaveOrder;