+++ jabgui/src/main/java/org/jabref/gui/util/BaseDialog.java
@@ -21,8 +21,7 @@ import javafx.stage.WindowEvent;
 import org.jabref.gui.icon.IconTheme;
 import org.jabref.gui.keyboard.KeyBinding;
 import org.jabref.gui.keyboard.KeyBindingRepository;
-
-import com.airhacks.afterburner.injection.Injector;
+import org.jabref.injection.Injector;
 
 public class BaseDialog<T> extends Dialog<T> {