+++ jabkit/src/main/java/module-info.java
@@ -17,7 +17,6 @@ module org.jabref.jabkit {
     requires java.prefs;
 
     requires javafx.base;
-    requires afterburner.fx;
 
     requires org.slf4j;
     requires jul.to.slf4j;